+++ algorithms/src/binary_search.c
@@ -1,13 +1,12 @@
-#include <stdio.h>
-#include "binary_search.h"
+#include "../include/binary_search.h"
 
-int binarySearch(int array[], size_t size, int target) {
+int binarySearch(int array[], int size, int target) {
     int l = 0;
     int r = size - 1;
-    
+
     while (l <= r) {
         int m = (l + r) / 2;
-        
+
         if (array[m] == target)
             return m;
         else if (array[m] < target)
@@ -15,6 +14,6 @@ int binarySearch(int array[], size_t size, int target) {
         else
             r = m - 1;
     }
-    
+
     return -1;
 }
